feat: :sparkles: Added Video Player
The following was done for the video player: - A path was defined to access the video player, this requires the id of a video. - A screen for the video player was created. In it the VideoPlayerBloc is instantiated to which VideoPlayerRepositoryImpl is injected. - VideoPlayerBloc has several states for handling video loading and capturing. - VideoPlayerBloc makes a request to the backend with the video id and the authentication token to obtain the video with the src (fetching). - Then, in VideoFullScreen the initialization of the video is performed (video downloading) - The VideoPlayer library performs the video playback. - VideoPlayBloc uses the VideoPlayerService to control the video (pause, resume, mute). - VIdeoButtons contains the play buttons, progress bar, mute button, etc... It makes use of the VideoPlayerBloc, so it is not coupled to the implementation details of the video playback. Aspects to keep in mind: - The authentication token is hardcoded, it must be extracted from a local storage. - It is not yet possible to fast forward or skip a video.   - The progress bar is the one that detects when a video is finished. - When the video buttons disappear, they still exist on the screen even if they are not visible, so you can interact with them.